Added script to .tmux.conf so a new split pane will inherit the Conda env from the pane it was split from.
The script will set a new Environment Variable, TMUX_NEW_CONDA_ENV, in
the new pane, and then the zshrc will read that variable and `source
activate` the right env.

It will inherit the correct environment from the pane it was split from.
That is, you can have multiple panes each with a different conda env,
and it will `activate` the same env that was active in the pane from
which the split command was executed.

Shortcomings:
 - This only works for my *mapped* keybindings (`%` and `"`) to split
   panes, it does not work with `:split-window`.
 - It is very specific to conda, and is not general to other settings.

Other notes:
 - I set it in my zshrc, not my .bashrc, just because i don't actually
   have my bashrc under source control right now. But this means it
   won't work if tmux is opening bash windows.
 - One cool side-effect is that if you `export TMUX_NEW_CONDA_ENV=myenv`
   before opening a new tmux-session, all the windows & panes in that
   session will activate the correct conda environment, meaning you
   don't have to temporarily modify your bashrc as is recommended here,
   for example:
    https://github.com/openai/universe-starter-agent#getting-started
        """ Add the following to your `.bashrc` so that you'll have the
        correct environment when the `train.py` script spawns new bash
        shells: `source activate universe-starter-agent` """